requirements.txt
numpy~=1.22.4
PyQt5~=5.15.6
Pillow~=9.1.1
scipy~=1.8.1
scikit-image~=0.19.2
tqdm~=4.64.0